Move the majority of the ElectrumX code
Much remains to be done but the demarcation is now clear.
To do is moving ElectrumX specific attributes to the electrumx module and
making the network configurable somehow.
